migration imports to new DB structure
