--- mealie/services/backups/import.py
@@ -1,9 +1,0 @@
-import json
-import shutil
-import zipfile
-from pathlib import Path
-
-from utils.logger import logger
-
-from settings import IMG_DIR, BACKUP_DIR, TEMPLATE_DIR, TEMP_DIR
-
